refactor: NumberIncrasementToggle => NumberIncrementer 이름 변경
